Refactor: modularize parser into domain-specific rule clases
