+++ src/parser/parser_utils.py
@@ -1,0 +1,3 @@
+def _pos(p, i) -> tuple[int, int]:
+    """Returns approximate (line, col) for the i-th symbol of the production."""
+    return p.lineno(i), p.lexpos(i)
