+++ src/parser/parsetab.py
@@ -6,9 +6,9 @@ _tabversion = '3.10'
 
 _lr_method = 'LALR'
 
-_lr_signature = 'moduleleftORleftANDleftEQUALSNOT_EQUALSLESS_THANLESS_THAN_EQUALSGREATER_THANGREATER_THAN_EQUALSleftPLUSMINUSleftTIMESDIVIDEFLOOR_DIVIDEMODrightUPLUSUMINUSNOTrightPOWERAND ASSIGN BREAK CLASS COLON COMMA CONTINUE DEDENT DEF DIVIDE DIVIDE_ASSIGN DOT ELIF ELSE EQUALS FALSE FLOOR_DIVIDE FLOOR_DIVIDE_ASSIGN FOR GREATER_THAN GREATER_THAN_EQUALS ID IF IN INDENT LBRACE LBRACKET LESS_THAN LESS_THAN_EQUALS LPAREN MINUS MINUS_ASSIGN MOD MOD_ASSIGN NEWLINE NONE NOT NOT_EQUALS NUMBER OR PASS PLUS PLUS_ASSIGN POWER POWER_ASSIGN RBRACE RBRACKET RETURN RPAREN STRING TIMES TIMES_ASSIGN TRUE WHILEmodule : statement_liststatement_list : statement\n| statement_list statementstatement : simple_statement\n| compound_statementsimple_statement : small_stmtsmall_stmt : assignment\n| return_stmt\n| break_stmt\n| continue_stmt\n| pass_stmt\n| exprexpr : atomexpr : LPAREN expr RPARENatom : NUMBERatom : STRINGatom : TRUEatom : FALSEatom : NONEatom : IDatom : LPAREN elements_opt RPARENatom : LBRACKET elements_opt RBRACKETatom : LBRACE key_value_list_opt RBRACEexpr : PLUS expr %prec UPLUSexpr : MINUS expr %prec UMINUSexpr : NOT exprexpr : expr POWER expr %prec POWERexpr : expr TIMES expr\n| expr DIVIDE expr\n| expr FLOOR_DIVIDE expr\n| expr MOD exprexpr : expr PLUS expr\n| expr MINUS exprexpr : expr EQUALS expr\n| expr NOT_EQUALS expr\n| expr LESS_THAN expr\n| expr LESS_THAN_EQUALS expr\n| expr GREATER_THAN expr\n| expr GREATER_THAN_EQUALS exprexpr : expr AND expr\n| expr OR exprexpr : ID LPAREN arg_list_opt RPARENarg_list_opt : expr COMMA arg_list_opt\n| exprarg_list_opt :empty :key_value_list : key_value\n| key_value COMMA key_value_listkey_value_list_opt : key_value_list\n| emptykey_value : expr COLON exprelements : expr\n| elements COMMA exprelements_opt : elements\n| emptyassignment : assign_targets ASSIGN expr\n| assign_targets PLUS_ASSIGN expr\n| assign_targets MINUS_ASSIGN expr\n| assign_targets TIMES_ASSIGN expr\n| assign_targets DIVIDE_ASSIGN expr\n| assign_targets FLOOR_DIVIDE_ASSIGN expr\n| assign_targets MOD_ASSIGN expr\n| assign_targets POWER_ASSIGN exprassign_targets : assign_targets ASSIGN target\n| targettarget : ID\n| LPAREN elements_opt RPAREN\n| LBRACKET elements_opt RBRACKETreturn_stmt : RETURN expr\n| RETURNbreak_stmt : BREAKcontinue_stmt : CONTINUEpass_stmt : PASScompound_statement : if_stmt\n| while_stmt\n| for_stmt\n| funcdef\n| classdefif_stmt : IF expr COLON suite elif_blocks else_block_optelif_blocks : ELIF expr COLON suite elif_blocks\n|else_block_opt : ELSE COLON suite\n|while_stmt : WHILE expr COLON suitefor_stmt : FOR ID IN expr COLON suitesuite : simple_statement\n| INDENT statement_list DEDENTfuncdef : DEF ID LPAREN param_list_opt RPAREN COLON statement_listclassdef : CLASS ID COLON statement_listparam_list_opt : param COMMA param_list_opt\n| param\n|param : ID'
+_lr_signature = 'moduleleftORleftANDleftEQUALSNOT_EQUALSLESS_THANLESS_THAN_EQUALSGREATER_THANGREATER_THAN_EQUALSleftPLUSMINUSleftTIMESDIVIDEFLOOR_DIVIDEMODrightUPLUSUMINUSNOTrightPOWERAND ASSIGN BREAK CLASS COLON COMMA CONTINUE DEDENT DEF DIVIDE DIVIDE_ASSIGN DOT ELIF ELSE EQUALS FALSE FLOOR_DIVIDE FLOOR_DIVIDE_ASSIGN FOR GREATER_THAN GREATER_THAN_EQUALS ID IF IN INDENT LBRACE LBRACKET LESS_THAN LESS_THAN_EQUALS LPAREN MINUS MINUS_ASSIGN MOD MOD_ASSIGN NEWLINE NONE NOT NOT_EQUALS NUMBER OR PASS PLUS PLUS_ASSIGN POWER POWER_ASSIGN RBRACE RBRACKET RETURN RPAREN STRING TIMES TIMES_ASSIGN TRUE WHILEcompound_statement : if_stmt\n| while_stmt\n| for_stmt\n| funcdef\n| classdefif_stmt : IF expr COLON suite elif_blocks else_block_optfuncdef : DEF ID LPAREN param_list_opt RPAREN COLON statement_listwhile_stmt : WHILE expr COLON suitestatement : simple_statement\n| compound_statementelif_blocks : ELIF expr COLON suite elif_blocks\n|expr : atomfor_stmt : FOR ID IN expr COLON suitesimple_statement : small_stmtclassdef : CLASS ID COLON statement_listsuite : simple_statement\n| INDENT statement_list DEDENTexpr : LPAREN expr RPARENsmall_stmt : assignment\n| return_stmt\n| break_stmt\n| continue_stmt\n| pass_stmt\n| exprelse_block_opt : ELSE COLON suite\n|param_list_opt : param COMMA param_list_opt\n| param\n|atom : NUMBERatom : STRINGparam : IDatom : TRUEassignment : assign_targets ASSIGN expr\n| assign_targets PLUS_ASSIGN expr\n| assign_targets MINUS_ASSIGN expr\n| assign_targets TIMES_ASSIGN expr\n| assign_targets DIVIDE_ASSIGN expr\n| assign_targets FLOOR_DIVIDE_ASSIGN expr\n| assign_targets MOD_ASSIGN expr\n| assign_targets POWER_ASSIGN expratom : FALSEatom : NONEatom : IDassign_targets : assign_targets ASSIGN target\n| targetatom : LPAREN elements_opt RPARENtarget : ID\n| LPAREN elements_opt RPAREN\n| LBRACKET elements_opt RBRACKETatom : LBRACKET elements_opt RBRACKETatom : LBRACE key_value_list_opt RBRACEreturn_stmt : RETURN expr\n| RETURNexpr : PLUS expr %prec UPLUSexpr : MINUS expr %prec UMINUSbreak_stmt : BREAKexpr : NOT exprcontinue_stmt : CONTINUEpass_stmt : PASSexpr : expr POWER expr %prec POWERmodule : statement_listexpr : expr TIMES expr\n| expr DIVIDE expr\n| expr FLOOR_DIVIDE expr\n| expr MOD exprstatement_list : statement\n| statement_list statementexpr : expr PLUS expr\n| expr MINUS exprexpr : expr EQUALS expr\n| expr NOT_EQUALS expr\n| expr LESS_THAN expr\n| expr LESS_THAN_EQUALS expr\n| expr GREATER_THAN expr\n| expr GREATER_THAN_EQUALS exprexpr : expr AND expr\n| expr OR exprexpr : ID LPAREN arg_list_opt RPARENarg_list_opt : expr COMMA arg_list_opt\n| exprarg_list_opt :key_value_list : key_value\n| key_value COMMA key_value_listkey_value_list_opt : key_value_list\n| emptykey_value : expr COLON exprelements : expr\n| elements COMMA exprelements_opt : elements\n| emptyempty :'
     
-_lr_action_items = {'IF':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[18,18,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-21,18,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,18,-21,-22,-84,-42,18,-83,18,-21,-22,-79,-87,-85,18,18,-82,-81,-80,]),'WHILE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[19,19,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-21,19,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,19,-21,-22,-84,-42,19,-83,19,-21,-22,-79,-87,-85,19,19,-82,-81,-80,]),'FOR':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[20,20,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-21,20,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,20,-21,-22,-84,-42,20,-83,20,-21,-22,-79,-87,-85,20,20,-82,-81,-80,]),'DEF':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[22,22,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-21,22,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,22,-21,-22,-84,-42,22,-83,22,-21,-22,-79,-87,-85,22,22,-82,-81,-80,]),'CLASS':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[24,24,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-21,24,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,24,-21,-22,-84,-42,24,-83,24,-21,-22,-79,-87,-85,24,24,-82,-81,-80,]),'RETURN':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[26,26,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,26,26,-14,-21,26,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,26,-21,-22,-84,-42,26,-83,26,26,-21,-22,-79,-87,-85,26,26,26,26,-82,-81,-80,]),'BREAK':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[27,27,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,27,27,-14,-21,27,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,27,-21,-22,-84,-42,27,-83,27,27,-21,-22,-79,-87,-85,27,27,27,27,-82,-81,-80,]),'CONTINUE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[28,28,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,28,28,-14,-21,28,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,28,-21,-22,-84,-42,28,-83,28,28,-21,-22,-79,-87,-85,28,28,28,28,-82,-81,-80,]),'PASS':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[29,29,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,29,29,-14,-21,29,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,29,-21,-22,-84,-42,29,-83,29,29,-21,-22,-79,-87,-85,29,29,29,29,-82,-81,-80,]),'LPAREN':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,65,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[23,23,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,59,59,64,59,59,-71,-72,-73,-13,59,59,59,-15,-16,-17,-18,-19,59,59,-3,59,59,59,59,59,59,59,59,59,59,59,59,59,59,59,59,64,59,59,112,119,59,59,59,59,59,59,59,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,23,23,59,-14,-21,59,23,-56,59,64,59,-57,-58,-59,-60,-61,-62,-63,-22,-23,59,59,-81,-86,23,-21,-22,-84,-42,59,23,-83,59,23,23,-21,-22,-79,-87,-85,23,23,23,23,-82,-81,-80,]),'PLUS':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,64,66,71,72,73,74,75,76,77,78,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,111,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,145,146,150,151,152,153,154,158,159,160,162,163,164,165,167,168,169,170,171,172,],[31,31,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,48,31,31,-20,31,31,-71,-72,-73,-13,31,31,31,-15,-16,-17,-18,-19,31,31,-3,31,31,31,31,31,31,31,31,31,31,31,31,31,31,31,48,31,-20,31,48,31,48,31,31,31,31,31,31,31,31,48,-24,-25,-26,48,48,-27,-28,-29,-30,-31,-32,-33,48,48,48,48,48,48,48,48,31,31,31,48,-14,-21,31,31,48,31,-20,31,48,48,48,48,48,48,48,-22,-23,31,31,-81,-86,31,-21,-22,-84,48,-42,31,48,31,48,-83,31,31,31,-21,-22,-79,48,-87,-85,31,31,31,31,-82,-81,-80,]),'MINUS':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,64,66,71,72,73,74,75,76,77,78,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,111,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,145,146,150,151,152,153,154,158,159,160,162,163,164,165,167,168,169,170,171,172,],[32,32,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,49,32,32,-20,32,32,-71,-72,-73,-13,32,32,32,-15,-16,-17,-18,-19,32,32,-3,32,32,32,32,32,32,32,32,32,32,32,32,32,32,32,49,32,-20,32,49,32,49,32,32,32,32,32,32,32,32,49,-24,-25,-26,49,49,-27,-28,-29,-30,-31,-32,-33,49,49,49,49,49,49,49,49,32,32,32,49,-14,-21,32,32,49,32,-20,32,49,49,49,49,49,49,49,-22,-23,32,32,-81,-86,32,-21,-22,-84,49,-42,32,49,32,49,-83,32,32,32,-21,-22,-79,49,-87,-85,32,32,32,32,-82,-81,-80,]),'NOT':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[33,33,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,33,33,-20,33,33,-71,-72,-73,-13,33,33,33,-15,-16,-17,-18,-19,33,33,-3,33,33,33,33,33,33,33,33,33,33,33,33,33,33,33,33,-20,33,33,33,33,33,33,33,33,33,33,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,33,33,33,-14,-21,33,33,-56,33,-20,33,-57,-58,-59,-60,-61,-62,-63,-22,-23,33,33,-81,-86,33,-21,-22,-84,-42,33,33,-83,33,33,33,-21,-22,-79,-87,-85,33,33,33,33,-82,-81,-80,]),'ID':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,112,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,157,158,159,160,163,164,165,167,168,169,170,171,172,],[21,21,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,60,60,63,-20,65,60,70,60,-71,-72,-73,-13,60,60,60,-15,-16,-17,-18,-19,60,60,-3,60,60,60,60,60,60,60,60,60,60,60,60,60,60,60,60,-20,60,60,120,60,60,60,60,60,60,60,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,21,21,60,142,-14,-21,60,21,-56,60,-20,60,-57,-58,-59,-60,-61,-62,-63,-22,-23,60,60,-81,-86,21,-21,-22,-84,-42,60,21,-83,60,21,21,142,-21,-22,-79,-87,-85,21,21,21,21,-82,-81,-80,]),'NUMBER':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[35,35,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,35,35,-20,35,35,-71,-72,-73,-13,35,35,35,-15,-16,-17,-18,-19,35,35,-3,35,35,35,35,35,35,35,35,35,35,35,35,35,35,35,35,-20,35,35,35,35,35,35,35,35,35,35,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,35,35,35,-14,-21,35,35,-56,35,-20,35,-57,-58,-59,-60,-61,-62,-63,-22,-23,35,35,-81,-86,35,-21,-22,-84,-42,35,35,-83,35,35,35,-21,-22,-79,-87,-85,35,35,35,35,-82,-81,-80,]),'STRING':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[36,36,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,36,36,-20,36,36,-71,-72,-73,-13,36,36,36,-15,-16,-17,-18,-19,36,36,-3,36,36,36,36,36,36,36,36,36,36,36,36,36,36,36,36,-20,36,36,36,36,36,36,36,36,36,36,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,36,36,36,-14,-21,36,36,-56,36,-20,36,-57,-58,-59,-60,-61,-62,-63,-22,-23,36,36,-81,-86,36,-21,-22,-84,-42,36,36,-83,36,36,36,-21,-22,-79,-87,-85,36,36,36,36,-82,-81,-80,]),'TRUE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[37,37,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,37,37,-20,37,37,-71,-72,-73,-13,37,37,37,-15,-16,-17,-18,-19,37,37,-3,37,37,37,37,37,37,37,37,37,37,37,37,37,37,37,37,-20,37,37,37,37,37,37,37,37,37,37,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,37,37,37,-14,-21,37,37,-56,37,-20,37,-57,-58,-59,-60,-61,-62,-63,-22,-23,37,37,-81,-86,37,-21,-22,-84,-42,37,37,-83,37,37,37,-21,-22,-79,-87,-85,37,37,37,37,-82,-81,-80,]),'FALSE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[38,38,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,38,38,-20,38,38,-71,-72,-73,-13,38,38,38,-15,-16,-17,-18,-19,38,38,-3,38,38,38,38,38,38,38,38,38,38,38,38,38,38,38,38,-20,38,38,38,38,38,38,38,38,38,38,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,38,38,38,-14,-21,38,38,-56,38,-20,38,-57,-58,-59,-60,-61,-62,-63,-22,-23,38,38,-81,-86,38,-21,-22,-84,-42,38,38,-83,38,38,38,-21,-22,-79,-87,-85,38,38,38,38,-82,-81,-80,]),'NONE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[39,39,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,39,39,-20,39,39,-71,-72,-73,-13,39,39,39,-15,-16,-17,-18,-19,39,39,-3,39,39,39,39,39,39,39,39,39,39,39,39,39,39,39,39,-20,39,39,39,39,39,39,39,39,39,39,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,39,39,39,-14,-21,39,39,-56,39,-20,39,-57,-58,-59,-60,-61,-62,-63,-22,-23,39,39,-81,-86,39,-21,-22,-84,-42,39,39,-83,39,39,39,-21,-22,-79,-87,-85,39,39,39,39,-82,-81,-80,]),'LBRACKET':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[40,40,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,61,61,-20,61,61,-71,-72,-73,-13,61,61,61,-15,-16,-17,-18,-19,61,61,-3,61,61,61,61,61,61,61,61,61,61,61,61,61,61,61,61,-20,61,61,121,61,61,61,61,61,61,61,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,40,40,61,-14,-21,61,40,-56,61,-20,61,-57,-58,-59,-60,-61,-62,-63,-22,-23,61,61,-81,-86,40,-21,-22,-84,-42,61,40,-83,61,40,40,-21,-22,-79,-87,-85,40,40,40,40,-82,-81,-80,]),'LBRACE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[41,41,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,41,41,-20,41,41,-71,-72,-73,-13,41,41,41,-15,-16,-17,-18,-19,41,41,-3,41,41,41,41,41,41,41,41,41,41,41,41,41,41,41,41,-20,41,41,41,41,41,41,41,41,41,41,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,41,41,41,-14,-21,41,41,-56,41,-20,41,-57,-58,-59,-60,-61,-62,-63,-22,-23,41,41,-81,-86,41,-21,-22,-84,-42,41,41,-83,41,41,41,-21,-22,-79,-87,-85,41,41,41,41,-82,-81,-80,]),'$end':([1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,117,120,122,123,124,125,126,127,128,129,130,133,134,136,137,138,140,146,151,158,159,160,163,164,169,170,171,172,],[0,-1,-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-21,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,-21,-22,-84,-42,-89,-83,-21,-22,-79,-87,-85,-88,-82,-81,-80,]),'DEDENT':([3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,117,120,122,123,124,125,126,127,128,129,130,133,134,136,137,138,140,146,151,153,158,159,160,163,164,169,170,171,172,],[-2,-4,-5,-6,-74,-75,-76,-77,-78,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-3,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-21,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,-21,-22,-84,-42,-89,-83,163,-21,-22,-79,-87,-85,-88,-82,-81,-80,]),'ELIF':([6,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,117,120,122,123,124,125,126,127,128,129,130,133,134,136,137,140,158,159,163,171,],[-6,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-21,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,152,-86,-21,-22,-42,-21,-22,-87,152,]),'ELSE':([6,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,117,120,122,123,124,125,126,127,128,129,130,133,134,136,137,140,151,158,159,163,171,172,],[-6,-7,-8,-9,-10,-11,-12,-20,-70,-71,-72,-73,-13,-15,-16,-17,-18,-19,-20,-69,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-21,-56,-20,-57,-58,-59,-60,-61,-62,-63,-22,-23,-81,-86,-21,-22,-42,161,-21,-22,-87,-81,-80,]),'POWER':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[43,-20,-13,-15,-16,-17,-18,-19,43,-20,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,-14,-21,43,-20,43,43,43,43,43,43,43,-22,-23,-21,-22,43,-42,43,43,-21,-22,43,]),'TIMES':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[44,-20,-13,-15,-16,-17,-18,-19,44,-20,44,44,44,-24,-25,-26,44,44,-27,-28,-29,-30,-31,44,44,44,44,44,44,44,44,44,44,44,-14,-21,44,-20,44,44,44,44,44,44,44,-22,-23,-21,-22,44,-42,44,44,-21,-22,44,]),'DIVIDE':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[45,-20,-13,-15,-16,-17,-18,-19,45,-20,45,45,45,-24,-25,-26,45,45,-27,-28,-29,-30,-31,45,45,45,45,45,45,45,45,45,45,45,-14,-21,45,-20,45,45,45,45,45,45,45,-22,-23,-21,-22,45,-42,45,45,-21,-22,45,]),'FLOOR_DIVIDE':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[46,-20,-13,-15,-16,-17,-18,-19,46,-20,46,46,46,-24,-25,-26,46,46,-27,-28,-29,-30,-31,46,46,46,46,46,46,46,46,46,46,46,-14,-21,46,-20,46,46,46,46,46,46,46,-22,-23,-21,-22,46,-42,46,46,-21,-22,46,]),'MOD':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[47,-20,-13,-15,-16,-17,-18,-19,47,-20,47,47,47,-24,-25,-26,47,47,-27,-28,-29,-30,-31,47,47,47,47,47,47,47,47,47,47,47,-14,-21,47,-20,47,47,47,47,47,47,47,-22,-23,-21,-22,47,-42,47,47,-21,-22,47,]),'EQUALS':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[50,-20,-13,-15,-16,-17,-18,-19,50,-20,50,50,50,-24,-25,-26,50,50,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,50,50,50,-14,-21,50,-20,50,50,50,50,50,50,50,-22,-23,-21,-22,50,-42,50,50,-21,-22,50,]),'NOT_EQUALS':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[51,-20,-13,-15,-16,-17,-18,-19,51,-20,51,51,51,-24,-25,-26,51,51,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,51,51,51,-14,-21,51,-20,51,51,51,51,51,51,51,-22,-23,-21,-22,51,-42,51,51,-21,-22,51,]),'LESS_THAN':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[52,-20,-13,-15,-16,-17,-18,-19,52,-20,52,52,52,-24,-25,-26,52,52,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,52,52,52,-14,-21,52,-20,52,52,52,52,52,52,52,-22,-23,-21,-22,52,-42,52,52,-21,-22,52,]),'LESS_THAN_EQUALS':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[53,-20,-13,-15,-16,-17,-18,-19,53,-20,53,53,53,-24,-25,-26,53,53,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,53,53,53,-14,-21,53,-20,53,53,53,53,53,53,53,-22,-23,-21,-22,53,-42,53,53,-21,-22,53,]),'GREATER_THAN':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[54,-20,-13,-15,-16,-17,-18,-19,54,-20,54,54,54,-24,-25,-26,54,54,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,54,54,54,-14,-21,54,-20,54,54,54,54,54,54,54,-22,-23,-21,-22,54,-42,54,54,-21,-22,54,]),'GREATER_THAN_EQUALS':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[55,-20,-13,-15,-16,-17,-18,-19,55,-20,55,55,55,-24,-25,-26,55,55,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,55,55,55,-14,-21,55,-20,55,55,55,55,55,55,55,-22,-23,-21,-22,55,-42,55,55,-21,-22,55,]),'AND':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[56,-20,-13,-15,-16,-17,-18,-19,56,-20,56,56,56,-24,-25,-26,56,56,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,56,56,-14,-21,56,-20,56,56,56,56,56,56,56,-22,-23,-21,-22,56,-42,56,56,-21,-22,56,]),'OR':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[57,-20,-13,-15,-16,-17,-18,-19,57,-20,57,57,57,-24,-25,-26,57,57,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,57,-14,-21,57,-20,57,57,57,57,57,57,57,-22,-23,-21,-22,57,-42,57,57,-21,-22,57,]),'ASSIGN':([21,25,34,114,118,120,129,158,159,],[-66,71,-65,-67,-64,-66,-68,-67,-68,]),'PLUS_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-66,72,-65,-67,-64,-66,-68,-67,-68,]),'MINUS_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-66,73,-65,-67,-64,-66,-68,-67,-68,]),'TIMES_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-66,74,-65,-67,-64,-66,-68,-67,-68,]),'DIVIDE_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-66,75,-65,-67,-64,-66,-68,-67,-68,]),'FLOOR_DIVIDE_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-66,76,-65,-67,-64,-66,-68,-67,-68,]),'MOD_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-66,77,-65,-67,-64,-66,-68,-67,-68,]),'POWER_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-66,78,-65,-67,-64,-66,-68,-67,-68,]),'RPAREN':([23,30,35,36,37,38,39,59,60,64,66,67,68,69,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,106,110,111,112,113,119,130,136,137,140,141,142,143,144,145,147,155,157,166,],[-46,-13,-15,-16,-17,-18,-19,-46,-20,-45,113,114,-54,-55,-24,-25,-26,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,136,140,-44,-92,-14,-46,-23,-21,-22,-42,-45,-93,156,-91,-53,158,-43,-92,-90,]),'COLON':([30,35,36,37,38,39,58,60,62,70,80,81,82,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,130,136,137,139,140,156,161,162,],[-13,-15,-16,-17,-18,-19,105,-20,108,116,-24,-25,-26,132,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-23,-21,-22,154,-42,165,167,168,]),'COMMA':([30,35,36,37,38,39,60,66,68,80,81,82,84,88,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,130,136,137,140,142,144,145,150,],[-13,-15,-16,-17,-18,-19,-20,-52,115,-24,-25,-26,-52,131,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,141,-14,-23,-21,-22,-42,-93,157,-53,-51,]),'RBRACKET':([30,35,36,37,38,39,40,60,61,68,69,80,81,82,83,84,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,107,113,121,130,136,137,140,145,148,],[-13,-15,-16,-17,-18,-19,-46,-20,-46,-54,-55,-24,-25,-26,129,-52,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,137,-14,-46,-23,-21,-22,-42,-53,159,]),'RBRACE':([30,35,36,37,38,39,41,60,80,81,82,85,86,87,88,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,130,136,137,140,149,150,],[-13,-15,-16,-17,-18,-19,-46,-20,-24,-25,-26,130,-49,-50,-47,-27,-28,-29,-30,-31,-32,-33,-34,-35,-36,-37,-38,-39,-40,-41,-14,-23,-21,-22,-42,-48,-51,]),'IN':([63,],[109,]),'INDENT':([105,108,154,167,168,],[135,135,135,135,135,]),}
+_lr_action_items = {'IF':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[18,18,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-48,18,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,18,-48,-52,-8,-80,18,-27,18,-48,-52,-6,-18,-14,18,18,-26,-12,-11,]),'WHILE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[19,19,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-48,19,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,19,-48,-52,-8,-80,19,-27,19,-48,-52,-6,-18,-14,19,19,-26,-12,-11,]),'FOR':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[20,20,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-48,20,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,20,-48,-52,-8,-80,20,-27,20,-48,-52,-6,-18,-14,20,20,-26,-12,-11,]),'DEF':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[22,22,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-48,22,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,22,-48,-52,-8,-80,22,-27,22,-48,-52,-6,-18,-14,22,22,-26,-12,-11,]),'CLASS':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,158,159,160,163,164,165,169,170,171,172,],[24,24,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-48,24,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,24,-48,-52,-8,-80,24,-27,24,-48,-52,-6,-18,-14,24,24,-26,-12,-11,]),'RETURN':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[26,26,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,26,26,-19,-48,26,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,26,-48,-52,-8,-80,26,-27,26,26,-48,-52,-6,-18,-14,26,26,26,26,-26,-12,-11,]),'BREAK':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[27,27,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,27,27,-19,-48,27,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,27,-48,-52,-8,-80,27,-27,27,27,-48,-52,-6,-18,-14,27,27,27,27,-26,-12,-11,]),'CONTINUE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[28,28,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,28,28,-19,-48,28,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,28,-48,-52,-8,-80,28,-27,28,28,-48,-52,-6,-18,-14,28,28,28,28,-26,-12,-11,]),'PASS':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,113,114,116,117,120,122,123,124,125,126,127,128,129,130,133,134,135,136,137,138,140,146,151,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[29,29,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,29,29,-19,-48,29,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,29,-48,-52,-8,-80,29,-27,29,29,-48,-52,-6,-18,-14,29,29,29,29,-26,-12,-11,]),'LPAREN':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,65,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[23,23,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,59,59,64,59,59,-58,-60,-61,-13,59,59,59,-31,-32,-34,-43,-44,59,59,-69,59,59,59,59,59,59,59,59,59,59,59,59,59,59,59,59,64,59,59,112,119,59,59,59,59,59,59,59,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,23,23,59,-19,-48,59,23,-35,59,64,59,-36,-37,-38,-39,-40,-41,-42,-52,-53,59,59,-12,-17,23,-48,-52,-8,-80,59,23,-27,59,23,23,-48,-52,-6,-18,-14,23,23,23,23,-26,-12,-11,]),'PLUS':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,64,66,71,72,73,74,75,76,77,78,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,111,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,145,146,150,151,152,153,154,158,159,160,162,163,164,165,167,168,169,170,171,172,],[31,31,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,48,31,31,-45,31,31,-58,-60,-61,-13,31,31,31,-31,-32,-34,-43,-44,31,31,-69,31,31,31,31,31,31,31,31,31,31,31,31,31,31,31,48,31,-45,31,48,31,48,31,31,31,31,31,31,31,31,48,-56,-57,-59,48,48,-62,-64,-65,-66,-67,-70,-71,48,48,48,48,48,48,48,48,31,31,31,48,-19,-48,31,31,48,31,-45,31,48,48,48,48,48,48,48,-52,-53,31,31,-12,-17,31,-48,-52,-8,48,-80,31,48,31,48,-27,31,31,31,-48,-52,-6,48,-18,-14,31,31,31,31,-26,-12,-11,]),'MINUS':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,58,59,60,61,62,64,66,71,72,73,74,75,76,77,78,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,111,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,139,140,141,145,146,150,151,152,153,154,158,159,160,162,163,164,165,167,168,169,170,171,172,],[32,32,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,49,32,32,-45,32,32,-58,-60,-61,-13,32,32,32,-31,-32,-34,-43,-44,32,32,-69,32,32,32,32,32,32,32,32,32,32,32,32,32,32,32,49,32,-45,32,49,32,49,32,32,32,32,32,32,32,32,49,-56,-57,-59,49,49,-62,-64,-65,-66,-67,-70,-71,49,49,49,49,49,49,49,49,32,32,32,49,-19,-48,32,32,49,32,-45,32,49,49,49,49,49,49,49,-52,-53,32,32,-12,-17,32,-48,-52,-8,49,-80,32,49,32,49,-27,32,32,32,-48,-52,-6,49,-18,-14,32,32,32,32,-26,-12,-11,]),'NOT':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[33,33,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,33,33,-45,33,33,-58,-60,-61,-13,33,33,33,-31,-32,-34,-43,-44,33,33,-69,33,33,33,33,33,33,33,33,33,33,33,33,33,33,33,33,-45,33,33,33,33,33,33,33,33,33,33,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,33,33,33,-19,-48,33,33,-35,33,-45,33,-36,-37,-38,-39,-40,-41,-42,-52,-53,33,33,-12,-17,33,-48,-52,-8,-80,33,33,-27,33,33,33,-48,-52,-6,-18,-14,33,33,33,33,-26,-12,-11,]),'ID':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,20,21,22,23,24,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,112,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,157,158,159,160,163,164,165,167,168,169,170,171,172,],[21,21,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,60,60,63,-45,65,60,70,60,-58,-60,-61,-13,60,60,60,-31,-32,-34,-43,-44,60,60,-69,60,60,60,60,60,60,60,60,60,60,60,60,60,60,60,60,-45,60,60,120,60,60,60,60,60,60,60,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,21,21,60,142,-19,-48,60,21,-35,60,-45,60,-36,-37,-38,-39,-40,-41,-42,-52,-53,60,60,-12,-17,21,-48,-52,-8,-80,60,21,-27,60,21,21,142,-48,-52,-6,-18,-14,21,21,21,21,-26,-12,-11,]),'NUMBER':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[35,35,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,35,35,-45,35,35,-58,-60,-61,-13,35,35,35,-31,-32,-34,-43,-44,35,35,-69,35,35,35,35,35,35,35,35,35,35,35,35,35,35,35,35,-45,35,35,35,35,35,35,35,35,35,35,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,35,35,35,-19,-48,35,35,-35,35,-45,35,-36,-37,-38,-39,-40,-41,-42,-52,-53,35,35,-12,-17,35,-48,-52,-8,-80,35,35,-27,35,35,35,-48,-52,-6,-18,-14,35,35,35,35,-26,-12,-11,]),'STRING':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[36,36,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,36,36,-45,36,36,-58,-60,-61,-13,36,36,36,-31,-32,-34,-43,-44,36,36,-69,36,36,36,36,36,36,36,36,36,36,36,36,36,36,36,36,-45,36,36,36,36,36,36,36,36,36,36,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,36,36,36,-19,-48,36,36,-35,36,-45,36,-36,-37,-38,-39,-40,-41,-42,-52,-53,36,36,-12,-17,36,-48,-52,-8,-80,36,36,-27,36,36,36,-48,-52,-6,-18,-14,36,36,36,36,-26,-12,-11,]),'TRUE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[37,37,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,37,37,-45,37,37,-58,-60,-61,-13,37,37,37,-31,-32,-34,-43,-44,37,37,-69,37,37,37,37,37,37,37,37,37,37,37,37,37,37,37,37,-45,37,37,37,37,37,37,37,37,37,37,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,37,37,37,-19,-48,37,37,-35,37,-45,37,-36,-37,-38,-39,-40,-41,-42,-52,-53,37,37,-12,-17,37,-48,-52,-8,-80,37,37,-27,37,37,37,-48,-52,-6,-18,-14,37,37,37,37,-26,-12,-11,]),'FALSE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[38,38,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,38,38,-45,38,38,-58,-60,-61,-13,38,38,38,-31,-32,-34,-43,-44,38,38,-69,38,38,38,38,38,38,38,38,38,38,38,38,38,38,38,38,-45,38,38,38,38,38,38,38,38,38,38,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,38,38,38,-19,-48,38,38,-35,38,-45,38,-36,-37,-38,-39,-40,-41,-42,-52,-53,38,38,-12,-17,38,-48,-52,-8,-80,38,38,-27,38,38,38,-48,-52,-6,-18,-14,38,38,38,38,-26,-12,-11,]),'NONE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[39,39,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,39,39,-45,39,39,-58,-60,-61,-13,39,39,39,-31,-32,-34,-43,-44,39,39,-69,39,39,39,39,39,39,39,39,39,39,39,39,39,39,39,39,-45,39,39,39,39,39,39,39,39,39,39,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,39,39,39,-19,-48,39,39,-35,39,-45,39,-36,-37,-38,-39,-40,-41,-42,-52,-53,39,39,-12,-17,39,-48,-52,-8,-80,39,39,-27,39,39,39,-48,-52,-6,-18,-14,39,39,39,39,-26,-12,-11,]),'LBRACKET':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[40,40,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,61,61,-45,61,61,-58,-60,-61,-13,61,61,61,-31,-32,-34,-43,-44,61,61,-69,61,61,61,61,61,61,61,61,61,61,61,61,61,61,61,61,-45,61,61,121,61,61,61,61,61,61,61,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,40,40,61,-19,-48,61,40,-35,61,-45,61,-36,-37,-38,-39,-40,-41,-42,-52,-53,61,61,-12,-17,40,-48,-52,-8,-80,61,40,-27,61,40,40,-48,-52,-6,-18,-14,40,40,40,40,-26,-12,-11,]),'LBRACE':([0,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18,19,21,23,26,27,28,29,30,31,32,33,35,36,37,38,39,40,41,42,43,44,45,46,47,48,49,50,51,52,53,54,55,56,57,59,60,61,64,71,72,73,74,75,76,77,78,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,105,108,109,113,114,115,116,117,119,120,121,122,123,124,125,126,127,128,129,130,131,132,133,134,135,136,137,138,140,141,146,151,152,153,154,158,159,160,163,164,165,167,168,169,170,171,172,],[41,41,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,41,41,-45,41,41,-58,-60,-61,-13,41,41,41,-31,-32,-34,-43,-44,41,41,-69,41,41,41,41,41,41,41,41,41,41,41,41,41,41,41,41,-45,41,41,41,41,41,41,41,41,41,41,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,41,41,41,-19,-48,41,41,-35,41,-45,41,-36,-37,-38,-39,-40,-41,-42,-52,-53,41,41,-12,-17,41,-48,-52,-8,-80,41,41,-27,41,41,41,-48,-52,-6,-18,-14,41,41,41,41,-26,-12,-11,]),'$end':([1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,117,120,122,123,124,125,126,127,128,129,130,133,134,136,137,138,140,146,151,158,159,160,163,164,169,170,171,172,],[0,-63,-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-48,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,-48,-52,-8,-80,-16,-27,-48,-52,-6,-18,-14,-7,-26,-12,-11,]),'DEDENT':([3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,42,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,117,120,122,123,124,125,126,127,128,129,130,133,134,136,137,138,140,146,151,153,158,159,160,163,164,169,170,171,172,],[-68,-9,-10,-15,-1,-2,-3,-4,-5,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-69,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-48,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,-48,-52,-8,-80,-16,-27,163,-48,-52,-6,-18,-14,-7,-26,-12,-11,]),'ELIF':([6,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,117,120,122,123,124,125,126,127,128,129,130,133,134,136,137,140,158,159,163,171,],[-15,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-48,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,152,-17,-48,-52,-80,-48,-52,-18,152,]),'ELSE':([6,12,13,14,15,16,17,21,26,27,28,29,30,35,36,37,38,39,60,79,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,114,117,120,122,123,124,125,126,127,128,129,130,133,134,136,137,140,151,158,159,163,171,172,],[-15,-20,-21,-22,-23,-24,-25,-45,-55,-58,-60,-61,-13,-31,-32,-34,-43,-44,-45,-54,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-48,-35,-45,-36,-37,-38,-39,-40,-41,-42,-52,-53,-12,-17,-48,-52,-80,161,-48,-52,-18,-12,-11,]),'POWER':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[43,-45,-13,-31,-32,-34,-43,-44,43,-45,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,43,-19,-48,43,-45,43,43,43,43,43,43,43,-52,-53,-48,-52,43,-80,43,43,-48,-52,43,]),'TIMES':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[44,-45,-13,-31,-32,-34,-43,-44,44,-45,44,44,44,-56,-57,-59,44,44,-62,-64,-65,-66,-67,44,44,44,44,44,44,44,44,44,44,44,-19,-48,44,-45,44,44,44,44,44,44,44,-52,-53,-48,-52,44,-80,44,44,-48,-52,44,]),'DIVIDE':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[45,-45,-13,-31,-32,-34,-43,-44,45,-45,45,45,45,-56,-57,-59,45,45,-62,-64,-65,-66,-67,45,45,45,45,45,45,45,45,45,45,45,-19,-48,45,-45,45,45,45,45,45,45,45,-52,-53,-48,-52,45,-80,45,45,-48,-52,45,]),'FLOOR_DIVIDE':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[46,-45,-13,-31,-32,-34,-43,-44,46,-45,46,46,46,-56,-57,-59,46,46,-62,-64,-65,-66,-67,46,46,46,46,46,46,46,46,46,46,46,-19,-48,46,-45,46,46,46,46,46,46,46,-52,-53,-48,-52,46,-80,46,46,-48,-52,46,]),'MOD':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[47,-45,-13,-31,-32,-34,-43,-44,47,-45,47,47,47,-56,-57,-59,47,47,-62,-64,-65,-66,-67,47,47,47,47,47,47,47,47,47,47,47,-19,-48,47,-45,47,47,47,47,47,47,47,-52,-53,-48,-52,47,-80,47,47,-48,-52,47,]),'EQUALS':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[50,-45,-13,-31,-32,-34,-43,-44,50,-45,50,50,50,-56,-57,-59,50,50,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,50,50,50,-19,-48,50,-45,50,50,50,50,50,50,50,-52,-53,-48,-52,50,-80,50,50,-48,-52,50,]),'NOT_EQUALS':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[51,-45,-13,-31,-32,-34,-43,-44,51,-45,51,51,51,-56,-57,-59,51,51,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,51,51,51,-19,-48,51,-45,51,51,51,51,51,51,51,-52,-53,-48,-52,51,-80,51,51,-48,-52,51,]),'LESS_THAN':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[52,-45,-13,-31,-32,-34,-43,-44,52,-45,52,52,52,-56,-57,-59,52,52,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,52,52,52,-19,-48,52,-45,52,52,52,52,52,52,52,-52,-53,-48,-52,52,-80,52,52,-48,-52,52,]),'LESS_THAN_EQUALS':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[53,-45,-13,-31,-32,-34,-43,-44,53,-45,53,53,53,-56,-57,-59,53,53,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,53,53,53,-19,-48,53,-45,53,53,53,53,53,53,53,-52,-53,-48,-52,53,-80,53,53,-48,-52,53,]),'GREATER_THAN':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[54,-45,-13,-31,-32,-34,-43,-44,54,-45,54,54,54,-56,-57,-59,54,54,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,54,54,54,-19,-48,54,-45,54,54,54,54,54,54,54,-52,-53,-48,-52,54,-80,54,54,-48,-52,54,]),'GREATER_THAN_EQUALS':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[55,-45,-13,-31,-32,-34,-43,-44,55,-45,55,55,55,-56,-57,-59,55,55,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,55,55,55,-19,-48,55,-45,55,55,55,55,55,55,55,-52,-53,-48,-52,55,-80,55,55,-48,-52,55,]),'AND':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[56,-45,-13,-31,-32,-34,-43,-44,56,-45,56,56,56,-56,-57,-59,56,56,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,56,56,-19,-48,56,-45,56,56,56,56,56,56,56,-52,-53,-48,-52,56,-80,56,56,-48,-52,56,]),'OR':([17,21,30,35,36,37,38,39,58,60,62,66,79,80,81,82,84,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,114,117,120,122,123,124,125,126,127,128,129,130,136,137,139,140,145,150,158,159,162,],[57,-45,-13,-31,-32,-34,-43,-44,57,-45,57,57,57,-56,-57,-59,57,57,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,57,-19,-48,57,-45,57,57,57,57,57,57,57,-52,-53,-48,-52,57,-80,57,57,-48,-52,57,]),'ASSIGN':([21,25,34,114,118,120,129,158,159,],[-49,71,-47,-50,-46,-49,-51,-50,-51,]),'PLUS_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-49,72,-47,-50,-46,-49,-51,-50,-51,]),'MINUS_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-49,73,-47,-50,-46,-49,-51,-50,-51,]),'TIMES_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-49,74,-47,-50,-46,-49,-51,-50,-51,]),'DIVIDE_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-49,75,-47,-50,-46,-49,-51,-50,-51,]),'FLOOR_DIVIDE_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-49,76,-47,-50,-46,-49,-51,-50,-51,]),'MOD_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-49,77,-47,-50,-46,-49,-51,-50,-51,]),'POWER_ASSIGN':([21,25,34,114,118,120,129,158,159,],[-49,78,-47,-50,-46,-49,-51,-50,-51,]),'RPAREN':([23,30,35,36,37,38,39,59,60,64,66,67,68,69,80,81,82,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,106,110,111,112,113,119,130,136,137,140,141,142,143,144,145,147,155,157,166,],[-93,-13,-31,-32,-34,-43,-44,-93,-45,-83,113,114,-91,-92,-56,-57,-59,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,136,140,-82,-30,-19,-93,-53,-48,-52,-80,-83,-33,156,-29,-90,158,-81,-30,-28,]),'COLON':([30,35,36,37,38,39,58,60,62,70,80,81,82,89,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,130,136,137,139,140,156,161,162,],[-13,-31,-32,-34,-43,-44,105,-45,108,116,-56,-57,-59,132,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-53,-48,-52,154,-80,165,167,168,]),'COMMA':([30,35,36,37,38,39,60,66,68,80,81,82,84,88,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,111,113,130,136,137,140,142,144,145,150,],[-13,-31,-32,-34,-43,-44,-45,-89,115,-56,-57,-59,-89,131,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,141,-19,-53,-48,-52,-80,-33,157,-90,-88,]),'RBRACKET':([30,35,36,37,38,39,40,60,61,68,69,80,81,82,83,84,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,107,113,121,130,136,137,140,145,148,],[-13,-31,-32,-34,-43,-44,-93,-45,-93,-91,-92,-56,-57,-59,129,-89,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,137,-19,-93,-53,-48,-52,-80,-90,159,]),'RBRACE':([30,35,36,37,38,39,41,60,80,81,82,85,86,87,88,90,91,92,93,94,95,96,97,98,99,100,101,102,103,104,113,130,136,137,140,149,150,],[-13,-31,-32,-34,-43,-44,-93,-45,-56,-57,-59,130,-86,-87,-84,-62,-64,-65,-66,-67,-70,-71,-72,-73,-74,-75,-76,-77,-78,-79,-19,-53,-48,-52,-80,-85,-88,]),'IN':([63,],[109,]),'INDENT':([105,108,154,167,168,],[135,135,135,135,135,]),}
 
 _lr_action = {}
 for _k, _v in _lr_action_items.items():
@@ -27,97 +27,97 @@ for _k, _v in _lr_goto_items.items():
 del _lr_goto_items
 _lr_productions = [
   ("S' -> module","S'",1,None,None,None),
-  ('module -> statement_list','module',1,'p_module','parser.py',113),
-  ('statement_list -> statement','statement_list',1,'p_statement_list','parser.py',121),
-  ('statement_list -> statement_list statement','statement_list',2,'p_statement_list','parser.py',122),
-  ('statement -> simple_statement','statement',1,'p_statement','parser.py',151),
-  ('statement -> compound_statement','statement',1,'p_statement','parser.py',152),
-  ('simple_statement -> small_stmt','simple_statement',1,'p_simple_statement','parser.py',157),
-  ('small_stmt -> assignment','small_stmt',1,'p_small_stmt','parser.py',163),
-  ('small_stmt -> return_stmt','small_stmt',1,'p_small_stmt','parser.py',164),
-  ('small_stmt -> break_stmt','small_stmt',1,'p_small_stmt','parser.py',165),
-  ('small_stmt -> continue_stmt','small_stmt',1,'p_small_stmt','parser.py',166),
-  ('small_stmt -> pass_stmt','small_stmt',1,'p_small_stmt','parser.py',167),
-  ('small_stmt -> expr','small_stmt',1,'p_small_stmt','parser.py',168),
-  ('expr -> atom','expr',1,'p_expr_atom','parser.py',182),
-  ('expr -> LPAREN expr RPAREN','expr',3,'p_expr_group','parser.py',188),
-  ('atom -> NUMBER','atom',1,'p_atom_number','parser.py',195),
-  ('atom -> STRING','atom',1,'p_atom_string','parser.py',200),
-  ('atom -> TRUE','atom',1,'p_atom_true','parser.py',205),
-  ('atom -> FALSE','atom',1,'p_atom_false','parser.py',210),
-  ('atom -> NONE','atom',1,'p_atom_none','parser.py',215),
-  ('atom -> ID','atom',1,'p_atom_identifier','parser.py',221),
-  ('atom -> LPAREN elements_opt RPAREN','atom',3,'p_atom_paren','parser.py',227),
-  ('atom -> LBRACKET elements_opt RBRACKET','atom',3,'p_atom_list','parser.py',239),
-  ('atom -> LBRACE key_value_list_opt RBRACE','atom',3,'p_atom_dict','parser.py',244),
-  ('expr -> PLUS expr','expr',2,'p_expr_unary_plus','parser.py',251),
-  ('expr -> MINUS expr','expr',2,'p_expr_unary_minus','parser.py',257),
-  ('expr -> NOT expr','expr',2,'p_expr_unary_not','parser.py',262),
-  ('expr -> expr POWER expr','expr',3,'p_expr_power','parser.py',269),
-  ('expr -> expr TIMES expr','expr',3,'p_expr_multiplicative','parser.py',275),
-  ('expr -> expr DIVIDE expr','expr',3,'p_expr_multiplicative','parser.py',276),
-  ('expr -> expr FLOOR_DIVIDE expr','expr',3,'p_expr_multiplicative','parser.py',277),
-  ('expr -> expr MOD expr','expr',3,'p_expr_multiplicative','parser.py',278),
-  ('expr -> expr PLUS expr','expr',3,'p_expr_additive','parser.py',284),
-  ('expr -> expr MINUS expr','expr',3,'p_expr_additive','parser.py',285),
-  ('expr -> expr EQUALS expr','expr',3,'p_expr_comparison','parser.py',293),
-  ('expr -> expr NOT_EQUALS expr','expr',3,'p_expr_comparison','parser.py',294),
-  ('expr -> expr LESS_THAN expr','expr',3,'p_expr_comparison','parser.py',295),
-  ('expr -> expr LESS_THAN_EQUALS expr','expr',3,'p_expr_comparison','parser.py',296),
-  ('expr -> expr GREATER_THAN expr','expr',3,'p_expr_comparison','parser.py',297),
-  ('expr -> expr GREATER_THAN_EQUALS expr','expr',3,'p_expr_comparison','parser.py',298),
-  ('expr -> expr AND expr','expr',3,'p_expr_logical','parser.py',306),
-  ('expr -> expr OR expr','expr',3,'p_expr_logical','parser.py',307),
-  ('expr -> ID LPAREN arg_list_opt RPAREN','expr',4,'p_expr_call','parser.py',315),
-  ('arg_list_opt -> expr COMMA arg_list_opt','arg_list_opt',3,'p_arg_list_opt','parser.py',328),
-  ('arg_list_opt -> expr','arg_list_opt',1,'p_arg_list_opt','parser.py',329),
-  ('arg_list_opt -> <empty>','arg_list_opt',0,'p_arg_list_opt_empty','parser.py',336),
-  ('empty -> <empty>','empty',0,'p_empty','parser.py',340),
-  ('key_value_list -> key_value','key_value_list',1,'p_key_value_list','parser.py',345),
-  ('key_value_list -> key_value COMMA key_value_list','key_value_list',3,'p_key_value_list','parser.py',346),
-  ('key_value_list_opt -> key_value_list','key_value_list_opt',1,'p_key_value_list_opt','parser.py',353),
-  ('key_value_list_opt -> empty','key_value_list_opt',1,'p_key_value_list_opt','parser.py',354),
-  ('key_value -> expr COLON expr','key_value',3,'p_key_value','parser.py',358),
-  ('elements -> expr','elements',1,'p_elements','parser.py',364),
-  ('elements -> elements COMMA expr','elements',3,'p_elements','parser.py',365),
-  ('elements_opt -> elements','elements_opt',1,'p_elements_opt','parser.py',372),
-  ('elements_opt -> empty','elements_opt',1,'p_elements_opt','parser.py',373),
-  ('assignment -> assign_targets ASSIGN expr','assignment',3,'p_assignment','parser.py',382),
-  ('assignment -> assign_targets PLUS_ASSIGN expr','assignment',3,'p_assignment','parser.py',383),
-  ('assignment -> assign_targets MINUS_ASSIGN expr','assignment',3,'p_assignment','parser.py',384),
-  ('assignment -> assign_targets TIMES_ASSIGN expr','assignment',3,'p_assignment','parser.py',385),
-  ('assignment -> assign_targets DIVIDE_ASSIGN expr','assignment',3,'p_assignment','parser.py',386),
-  ('assignment -> assign_targets FLOOR_DIVIDE_ASSIGN expr','assignment',3,'p_assignment','parser.py',387),
-  ('assignment -> assign_targets MOD_ASSIGN expr','assignment',3,'p_assignment','parser.py',388),
-  ('assignment -> assign_targets POWER_ASSIGN expr','assignment',3,'p_assignment','parser.py',389),
-  ('assign_targets -> assign_targets ASSIGN target','assign_targets',3,'p_assign_targets','parser.py',398),
-  ('assign_targets -> target','assign_targets',1,'p_assign_targets','parser.py',399),
-  ('target -> ID','target',1,'p_target','parser.py',406),
-  ('target -> LPAREN elements_opt RPAREN','target',3,'p_target','parser.py',407),
-  ('target -> LBRACKET elements_opt RBRACKET','target',3,'p_target','parser.py',408),
-  ('return_stmt -> RETURN expr','return_stmt',2,'p_return_stmt','parser.py',422),
-  ('return_stmt -> RETURN','return_stmt',1,'p_return_stmt','parser.py',423),
-  ('break_stmt -> BREAK','break_stmt',1,'p_break_stmt','parser.py',431),
-  ('continue_stmt -> CONTINUE','continue_stmt',1,'p_continue_stmt','parser.py',436),
-  ('pass_stmt -> PASS','pass_stmt',1,'p_pass_stmt','parser.py',441),
-  ('compound_statement -> if_stmt','compound_statement',1,'p_compound_statement','parser.py',447),
-  ('compound_statement -> while_stmt','compound_statement',1,'p_compound_statement','parser.py',448),
-  ('compound_statement -> for_stmt','compound_statement',1,'p_compound_statement','parser.py',449),
-  ('compound_statement -> funcdef','compound_statement',1,'p_compound_statement','parser.py',450),
-  ('compound_statement -> classdef','compound_statement',1,'p_compound_statement','parser.py',451),
-  ('if_stmt -> IF expr COLON suite elif_blocks else_block_opt','if_stmt',6,'p_if_stmt','parser.py',458),
-  ('elif_blocks -> ELIF expr COLON suite elif_blocks','elif_blocks',5,'p_elif_blocks','parser.py',463),
-  ('elif_blocks -> <empty>','elif_blocks',0,'p_elif_blocks','parser.py',464),
-  ('else_block_opt -> ELSE COLON suite','else_block_opt',3,'p_else_block_opt','parser.py',472),
-  ('else_block_opt -> <empty>','else_block_opt',0,'p_else_block_opt','parser.py',473),
-  ('while_stmt -> WHILE expr COLON suite','while_stmt',4,'p_while_stmt','parser.py',479),
-  ('for_stmt -> FOR ID IN expr COLON suite','for_stmt',6,'p_for_stmt','parser.py',486),
-  ('suite -> simple_statement','suite',1,'p_suite','parser.py',497),
-  ('suite -> INDENT statement_list DEDENT','suite',3,'p_suite','parser.py',498),
-  ('funcdef -> DEF ID LPAREN param_list_opt RPAREN COLON statement_list','funcdef',7,'p_funcdef','parser.py',512),
-  ('classdef -> CLASS ID COLON statement_list','classdef',4,'p_classdef','parser.py',518),
-  ('param_list_opt -> param COMMA param_list_opt','param_list_opt',3,'p_param_list_opt','parser.py',524),
-  ('param_list_opt -> param','param_list_opt',1,'p_param_list_opt','parser.py',525),
-  ('param_list_opt -> <empty>','param_list_opt',0,'p_param_list_opt','parser.py',526),
-  ('param -> ID','param',1,'p_param','parser.py',535),
+  ('compound_statement -> if_stmt','compound_statement',1,'p_compound_statement','parser_blocks.py',10),
+  ('compound_statement -> while_stmt','compound_statement',1,'p_compound_statement','parser_blocks.py',11),
+  ('compound_statement -> for_stmt','compound_statement',1,'p_compound_statement','parser_blocks.py',12),
+  ('compound_statement -> funcdef','compound_statement',1,'p_compound_statement','parser_blocks.py',13),
+  ('compound_statement -> classdef','compound_statement',1,'p_compound_statement','parser_blocks.py',14),
+  ('if_stmt -> IF expr COLON suite elif_blocks else_block_opt','if_stmt',6,'p_if_stmt','parser_conditionals.py',10),
+  ('funcdef -> DEF ID LPAREN param_list_opt RPAREN COLON statement_list','funcdef',7,'p_funcdef','parser_definitions.py',10),
+  ('while_stmt -> WHILE expr COLON suite','while_stmt',4,'p_while_stmt','parser_loops.py',10),
+  ('statement -> simple_statement','statement',1,'p_statement','parser_statements.py',10),
+  ('statement -> compound_statement','statement',1,'p_statement','parser_statements.py',11),
+  ('elif_blocks -> ELIF expr COLON suite elif_blocks','elif_blocks',5,'p_elif_blocks','parser_conditionals.py',14),
+  ('elif_blocks -> <empty>','elif_blocks',0,'p_elif_blocks','parser_conditionals.py',15),
+  ('expr -> atom','expr',1,'p_expr_atom','parser_expressions.py',14),
+  ('for_stmt -> FOR ID IN expr COLON suite','for_stmt',6,'p_for_stmt','parser_loops.py',14),
+  ('simple_statement -> small_stmt','simple_statement',1,'p_simple_statement','parser_statements.py',15),
+  ('classdef -> CLASS ID COLON statement_list','classdef',4,'p_classdef','parser_definitions.py',16),
+  ('suite -> simple_statement','suite',1,'p_suite','parser_blocks.py',18),
+  ('suite -> INDENT statement_list DEDENT','suite',3,'p_suite','parser_blocks.py',19),
+  ('expr -> LPAREN expr RPAREN','expr',3,'p_expr_group','parser_expressions.py',18),
+  ('small_stmt -> assignment','small_stmt',1,'p_small_stmt','parser_statements.py',19),
+  ('small_stmt -> return_stmt','small_stmt',1,'p_small_stmt','parser_statements.py',20),
+  ('small_stmt -> break_stmt','small_stmt',1,'p_small_stmt','parser_statements.py',21),
+  ('small_stmt -> continue_stmt','small_stmt',1,'p_small_stmt','parser_statements.py',22),
+  ('small_stmt -> pass_stmt','small_stmt',1,'p_small_stmt','parser_statements.py',23),
+  ('small_stmt -> expr','small_stmt',1,'p_small_stmt','parser_statements.py',24),
+  ('else_block_opt -> ELSE COLON suite','else_block_opt',3,'p_else_block_opt','parser_conditionals.py',22),
+  ('else_block_opt -> <empty>','else_block_opt',0,'p_else_block_opt','parser_conditionals.py',23),
+  ('param_list_opt -> param COMMA param_list_opt','param_list_opt',3,'p_param_list_opt','parser_definitions.py',22),
+  ('param_list_opt -> param','param_list_opt',1,'p_param_list_opt','parser_definitions.py',23),
+  ('param_list_opt -> <empty>','param_list_opt',0,'p_param_list_opt','parser_definitions.py',24),
+  ('atom -> NUMBER','atom',1,'p_atom_number','parser_expressions.py',23),
+  ('atom -> STRING','atom',1,'p_atom_string','parser_expressions.py',28),
+  ('param -> ID','param',1,'p_param','parser_definitions.py',33),
+  ('atom -> TRUE','atom',1,'p_atom_true','parser_expressions.py',33),
+  ('assignment -> assign_targets ASSIGN expr','assignment',3,'p_assignment','parser_statements.py',34),
+  ('assignment -> assign_targets PLUS_ASSIGN expr','assignment',3,'p_assignment','parser_statements.py',35),
+  ('assignment -> assign_targets MINUS_ASSIGN expr','assignment',3,'p_assignment','parser_statements.py',36),
+  ('assignment -> assign_targets TIMES_ASSIGN expr','assignment',3,'p_assignment','parser_statements.py',37),
+  ('assignment -> assign_targets DIVIDE_ASSIGN expr','assignment',3,'p_assignment','parser_statements.py',38),
+  ('assignment -> assign_targets FLOOR_DIVIDE_ASSIGN expr','assignment',3,'p_assignment','parser_statements.py',39),
+  ('assignment -> assign_targets MOD_ASSIGN expr','assignment',3,'p_assignment','parser_statements.py',40),
+  ('assignment -> assign_targets POWER_ASSIGN expr','assignment',3,'p_assignment','parser_statements.py',41),
+  ('atom -> FALSE','atom',1,'p_atom_false','parser_expressions.py',38),
+  ('atom -> NONE','atom',1,'p_atom_none','parser_expressions.py',43),
+  ('atom -> ID','atom',1,'p_atom_identifier','parser_expressions.py',48),
+  ('assign_targets -> assign_targets ASSIGN target','assign_targets',3,'p_assign_targets','parser_statements.py',49),
+  ('assign_targets -> target','assign_targets',1,'p_assign_targets','parser_statements.py',50),
+  ('atom -> LPAREN elements_opt RPAREN','atom',3,'p_atom_paren','parser_expressions.py',54),
+  ('target -> ID','target',1,'p_target','parser_statements.py',57),
+  ('target -> LPAREN elements_opt RPAREN','target',3,'p_target','parser_statements.py',58),
+  ('target -> LBRACKET elements_opt RBRACKET','target',3,'p_target','parser_statements.py',59),
+  ('atom -> LBRACKET elements_opt RBRACKET','atom',3,'p_atom_list','parser_expressions.py',63),
+  ('atom -> LBRACE key_value_list_opt RBRACE','atom',3,'p_atom_dict','parser_expressions.py',68),
+  ('return_stmt -> RETURN expr','return_stmt',2,'p_return_stmt','parser_statements.py',72),
+  ('return_stmt -> RETURN','return_stmt',1,'p_return_stmt','parser_statements.py',73),
+  ('expr -> PLUS expr','expr',2,'p_expr_unary_plus','parser_expressions.py',74),
+  ('expr -> MINUS expr','expr',2,'p_expr_unary_minus','parser_expressions.py',79),
+  ('break_stmt -> BREAK','break_stmt',1,'p_break_stmt','parser_statements.py',80),
+  ('expr -> NOT expr','expr',2,'p_expr_unary_not','parser_expressions.py',84),
+  ('continue_stmt -> CONTINUE','continue_stmt',1,'p_continue_stmt','parser_statements.py',84),
+  ('pass_stmt -> PASS','pass_stmt',1,'p_pass_stmt','parser_statements.py',88),
+  ('expr -> expr POWER expr','expr',3,'p_expr_power','parser_expressions.py',90),
+  ('module -> statement_list','module',1,'p_module','parser.py',93),
+  ('expr -> expr TIMES expr','expr',3,'p_expr_multiplicative','parser_expressions.py',95),
+  ('expr -> expr DIVIDE expr','expr',3,'p_expr_multiplicative','parser_expressions.py',96),
+  ('expr -> expr FLOOR_DIVIDE expr','expr',3,'p_expr_multiplicative','parser_expressions.py',97),
+  ('expr -> expr MOD expr','expr',3,'p_expr_multiplicative','parser_expressions.py',98),
+  ('statement_list -> statement','statement_list',1,'p_statement_list','parser.py',102),
+  ('statement_list -> statement_list statement','statement_list',2,'p_statement_list','parser.py',103),
+  ('expr -> expr PLUS expr','expr',3,'p_expr_additive','parser_expressions.py',103),
+  ('expr -> expr MINUS expr','expr',3,'p_expr_additive','parser_expressions.py',104),
+  ('expr -> expr EQUALS expr','expr',3,'p_expr_comparison','parser_expressions.py',110),
+  ('expr -> expr NOT_EQUALS expr','expr',3,'p_expr_comparison','parser_expressions.py',111),
+  ('expr -> expr LESS_THAN expr','expr',3,'p_expr_comparison','parser_expressions.py',112),
+  ('expr -> expr LESS_THAN_EQUALS expr','expr',3,'p_expr_comparison','parser_expressions.py',113),
+  ('expr -> expr GREATER_THAN expr','expr',3,'p_expr_comparison','parser_expressions.py',114),
+  ('expr -> expr GREATER_THAN_EQUALS expr','expr',3,'p_expr_comparison','parser_expressions.py',115),
+  ('expr -> expr AND expr','expr',3,'p_expr_logical','parser_expressions.py',121),
+  ('expr -> expr OR expr','expr',3,'p_expr_logical','parser_expressions.py',122),
+  ('expr -> ID LPAREN arg_list_opt RPAREN','expr',4,'p_expr_call','parser_expressions.py',128),
+  ('arg_list_opt -> expr COMMA arg_list_opt','arg_list_opt',3,'p_arg_list_opt','parser_expressions.py',139),
+  ('arg_list_opt -> expr','arg_list_opt',1,'p_arg_list_opt','parser_expressions.py',140),
+  ('arg_list_opt -> <empty>','arg_list_opt',0,'p_arg_list_opt_empty','parser_expressions.py',147),
+  ('key_value_list -> key_value','key_value_list',1,'p_key_value_list','parser_expressions.py',151),
+  ('key_value_list -> key_value COMMA key_value_list','key_value_list',3,'p_key_value_list','parser_expressions.py',152),
+  ('key_value_list_opt -> key_value_list','key_value_list_opt',1,'p_key_value_list_opt','parser_expressions.py',159),
+  ('key_value_list_opt -> empty','key_value_list_opt',1,'p_key_value_list_opt','parser_expressions.py',160),
+  ('key_value -> expr COLON expr','key_value',3,'p_key_value','parser_expressions.py',164),
+  ('elements -> expr','elements',1,'p_elements','parser_expressions.py',168),
+  ('elements -> elements COMMA expr','elements',3,'p_elements','parser_expressions.py',169),
+  ('elements_opt -> elements','elements_opt',1,'p_elements_opt','parser_expressions.py',176),
+  ('elements_opt -> empty','elements_opt',1,'p_elements_opt','parser_expressions.py',177),
+  ('empty -> <empty>','empty',0,'p_empty','parser_expressions.py',181),
 ]
