+++ src/testers/parser/test_statements.flpy
@@ -1,4 +1,4 @@
-# Asignaciones simples
+# Assignments
 a = 1
 b = 2
 c = 3
@@ -30,7 +30,7 @@ elif x == 3:
 else:
     result = 0
 
-# While con break y continue
+# While
 counter = 0
 while counter < 3:
     counter += 1
@@ -39,7 +39,7 @@ while counter < 3:
     if counter == 3:
         break
 
-# For loop simple
+# For loop
 for i in range(2):
     print(i)
